Move #load and #reload to separate module Loadable
