+++ lib/dynamoid.rb
@@ -30,6 +30,7 @@ require 'dynamoid/criteria'
 require 'dynamoid/finders'
 require 'dynamoid/identity_map'
 require 'dynamoid/config'
+require 'dynamoid/loadable'
 require 'dynamoid/components'
 require 'dynamoid/document'
 require 'dynamoid/adapter'
